parallelization unit test
